+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## 0.2.2(2023-09-19)
+- Refacto mutation
+- Working nested deletions and other cases that were not working
+- Simplification of the code
 ## 0.2.1(2023-09-04)
 - Disabled cardinality check as it has false positives
 ## 0.2.0(2023-08-11)